Revert "create/rename sections for questions"
This reverts commit 8325c8def5c7d9fae466d1f7a4f139c4f1684ecb.
